Fix MSVC /Wall /WX build for aie_codegen and apply BinSkim-oriented flags (#138)
* Fix MSVC /Wall /WX build for aie_codegen and apply BinSkim-oriented flags

Align the Windows build with strict MSVC warnings and BinSkim-oriented
settings: keep /Wall and /WX, move ASLR/CFG/CET/large-address flags from
compile options to link options, and resolve the resulting diagnostics in
runtime sources instead of silencing 4244/4018.

Code changes include explicit width for tile locations and DMA/register
fields, safer control-code file I/O and backward scans, correct fgetc
storage type, loop index typing for signed/unsigned comparisons, and
initializers where MSVC C4701 could not prove definite assignment.


* Improve MSVC build metadata and signed indexing in control-code / privilege paths

- Document MSVC compile options with a concise CMake comment instead of empty banners.
- Use u8 for the AIE partition clock column iterator and pass the column index to
  XAie_TileLoc without a redundant narrowing cast.
- In control-code DMA BD length updates, use ssize_t for backward indices, avoid
  mixed signed/unsigned loop and count checks, cast file offsets to long for
  fseek, snapshot ftell before seeking to end where required, and skip the scan
  when ftell fails.
- Drop an unnecessary compound statement around LoadCores cache label allocation
  now that C99/C11 mixed declarations are used elsewhere in the file.


* Harden MSVC builds for BinSkim and strict static analysis

Enable control-flow guard on both compile and link (/guard:cf) so BA2008
is satisfied when the product DLL is produced from this target. Move
BinSkim-oriented linker switches off the compiler (ASLR /dynamicbase,
/guard:cf, /largeaddressaware) and skip /CETCOMPAT when targeting VS ARM64,
where Intel CET does not apply.

Keep /Wall and /WX with a small set of /wd suppressions for Windows SDK and
third-party noise (including C5105 under older SDK + VS2019).

Add AIE_CODEGEN_MSVC_RELEASE_PDB (default ON) to emit Release /Zi and link
/debug so compliance tools (e.g. BinSkim BA2026) can better observe /sdl;
disable if Release PDBs must not be produced.

Non-MSVC: tighten parity with -Wall -Wextra -Werror.


---------

+++ src/device/xaie_device_aie2ps.c
@@ -246,7 +246,7 @@ AieRC _XAie2PS_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 
 	/*Enable the clock control register for shims*/
 	for(u32 C = Args->StartCol; C < (Args->StartCol + Args->NumCols); C++) {
-		XAie_LocType TileLoc = XAie_TileLoc(C, 1);
+		XAie_LocType TileLoc = XAie_TileLoc((u8)C, 1U);
 
 		RC = _XAie2PS_PmSetColumnClockBuffer(DevInst, TileLoc,
 				Args->Enable);
@@ -264,7 +264,8 @@ AieRC _XAie2PS_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 		}
 	}
 
-	TileStatus = _XAie_GetTileBitPosFromLoc(DevInst, XAie_TileLoc(Args->StartCol, 1));
+	TileStatus = _XAie_GetTileBitPosFromLoc(DevInst,
+			XAie_TileLoc((u8)Args->StartCol, 1U));
 	NumTiles =(u32)((DevInst->NumRows - 1U) * (Args->NumCols));
 
 	if(Args->Enable) {
